mingw gcc is in the path, pthreads missing
This successfully builds a shell with a mingw compiler, but the cargo
build now fails because it isn't able to link against pthreads.  This
seems like somewhat of a known (maybe solved) issue, but I can't figure
out how to incorporate the solution (yet).  Using the "buildPackages"
prefix also removes the requirement for `{allowUnsupporteSystem =
true;}`